Avax with prefix (#359)
* chore: package updates courtesy of npm

* chore: ignore vim swp files

* feat: support AVAX X- & P- chain IDs

AVAX uses bech32 for its X & P chains. A valid address format includes
the 'X-' or 'P-' ID prepended to the address. This adds a custom decoder
to strip the ID and decode the leftover address with bech32.

Change is backwards compatible with existing functionality.

* feat: Avalanche C-Chain support

The Avalanche C-Chain (AVAXC) uses the EVM addressing system. That being
said, on Avalanche, it is valid to have the chain ID prepended to the
address. This decoder does some preparsing to strip the ID and pass the
leftover address to a decoder.

* fix: linting rules

* Revert "chore: package updates courtesy of npm"

This reverts commit 8f74a3482600fc9a9373017b8205d90cd268c321.

* Remove avax-c implementation

Co-authored-by: charchar <87212793+0xcharchar@users.noreply.github.com>